Changed delay structure
The interval delay starts after the first byte is received, thus
making this function have the same problem of hanging the thread. The
total interval timeout is not much better, but it does set hard
deadlines. Using this type of delay the read function should be fixed
to be more careful about timeouts.